Added suport for SKIP_BUILD_DOCKER_IMAGE environment variable.
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@103742 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ docker/build-docker-image.py
@@ -36,10 +36,11 @@ set -e
 BUILD_NUMBER=${BUILD_NUMBER}
 NAME=%s
 
-function die() {
-  echo "ERROR: "$@
-  exit 1
-}
+eval skip=\$$SKIP_BUILD_DOCKER_IMAGE
+if [  -z "$skip" ]; then
+  echo skipping building of $NAME docker image
+  exit 0
+fi
 
 ARTIFACT=%s
 [ -e ${ARTIFACT} ] && rm ${ARTIFACT}
